backend/requirements.txt
django
django-cors-headers
djangorestframework
python-decouple
psycopg2-binary
doorstop